requirements.txt
streamlit
python-dotenv
requests
langchain
langchain-community
PyMuPDF
numpy
scikit-learn